Support deleted proofs
When a proof's deleted, it doesn't get sent to us in the proofs array;
it's not a real proof, that kind of info isn't available for it.  But
now (since CORE-2760) we get a revokedProofs array with some of the same
info in.  So the task of this PR Is to take revokedProofs and synthesize
Proofs out of its members, so that we can show a revoked proof inside
the popup.

Doing this uncovered two more bugs -- DESKTOP-756 and DESKTOP-757, the
first of which should be fixed before merging this one in.