Set myDeviceRole when starting login
We were using whether the service was "registered" to determine if
provisioning is happening with a NewComputer or ExistingComputer.
But someone could be registered under one username, and trying to log
in under a different one.  In that case, they should be a NewComputer,
not an ExistingComputer.

This PR sets myDeviceRole to NewComputer at the start of the login flow.