Add Native inputs for android & ios
This adds a new type for inputs called `passwordVisible`. currently not
implemented, but it should keep the platform from remembering what was
typed (for predictive typing) and be visible to the user. Currently it
just doesn't show the typing (default password behavior)

Quirks:
Text styles don't apply to android if you use a secure input.