Native inputs (#2535)
* Add Native inputs for android & ios

This adds a new type for inputs called `passwordVisible`. currently not
implemented, but it should keep the platform from remembering what was
typed (for predictive typing) and be visible to the user. Currently it
just doesn't show the typing (default password behavior)

Quirks:
Text styles don't apply to android if you use a secure input.

* Make Text Styles static functions so we can reuse

* Add hacky Visible Password on android

* Undo useless gradle wrapper change

* Update desktop inputs to use passwordVisible

* Implement onChangeText for desktop too

* Remove dead TODO

* Add multiline implementation. (#2553)

* Add multiline implmenetation.

I can't seem to figure out how to align the ios text vertically
downwards. Android does not have this problem.

* Remove dupes in stylesheet