folder_branch_ops: don't try to apply unreadable updates
Don't silently ignore the changes from MD updates that we couldn't
decrypt, since then the client will seem to be up-to-date with respect
to revision number, but really isn't.  Instead, error out while
applying the updates.

Also, when reading a batch of updates, try to re-decrypt any
unreadable updates after processing the entire batch.  The reason is
that the needed key might not be present until a later rekey update.