implement dismissals for show_tracker_popup gregor notifications
This creates a new `Dismiss()` function as part of the IdentifyUI
protocol, which Electron will need to implement. That RPC will tell
Electron to close existing track windows for a given user.

This PR implements receiving dismissals, but what remains to be done is
*sending* dismissals. This will require another new RPC, maybe called
`DismissWithToken()`, as a counterpart to `TrackWithToken()` in
`track.avdl`. This will be called by Electron when a track window is
dismissed without tracking. Both of those RPC will be used to send out
gregor dismissals in a subsequent PR.