avoid undefined behavior in Go assignments
The funky assignment that initialized p.Hash as part of the method on
the RHS, invoked undefined behavior in the Go compiler, and ended up
crashing when we built with gcc-go. There was a lot of dirty code there
anyway, so this diff just goes ahead and rewrites that whole bit.