libkbfs: avoid rekey loops on incomplete rekeys
If a reader handled a rekey request by setting the rekey bit, but
there was work it couldn't do on behalf of itself (like incrementing
the key generation due to a revoked device), it could enter into a
loop of continuous Put/SendRekeyNeeded messages back and forth with
the server.

Instead, detect if the rekeyer made any useful revisions, and if not,
don't bother putting a new revision.

Issue: KBFS-1119