Make tab content flex styled & views manage scroll (#3083)
* Make tab content flex styled & views manage scroll

Apply only display: flex and flex: 1 to containers in meta navigator
and tab-bar. This will expose a view to the child that is the full,
available screen size that they can play with, no extra overflow
attached. If the child needs to overflow, it can wrap it's contents in
a normal Box/Div that has scrollable: auto (provided as
globalStyles.scrollable). This separates the concerns of child
overflow management away from the tab bar and meta navigator and
provides more power to the children to maintain their scroll
behaviour/layout behaviour in a marginally impactful SLOC manner. This
also brings parity to how layouts are handled on native.

* Improve scrollable style and fix regressions on folders