try to parse client-out-of-date messages as JSON
This lets the server define a "cli_message" field, so that we can
differentiate between messages for the CLI and messages for the GUI. If
we fail to parse any step of this, we just fall back to printing the
entire message.

As part of this, factor out all this parsing and printing from main.go
to a new file in go/client.