Handle __SCREENSHOT__; consistant Avatar structure
Move __SCREENSHOT__ handling into avatar.shared so that it is ready for
the eventual migration of visdiff to mobile.

Take advantage of TouchableOpacity and TouchableWithoutHighlight to
always produce an Avatar view hierarchy that is the same depth. Use
disabled to keep touches from getting interecepted on Avatars without
click handlers.