folder_block_manager: don't clean up all blocks from failed sync
After a failed sync of a file, it's possible the next sync will use
the same blocks again, so don't necessarily clean them up (because the
clean up could happen after the successful Sync, leading to missing
blocks!).

Instead, when a sync fails and the next successful revision matches
the same merged status, clean up only those blocks that didn't make it
into the successful revision.  If the merged status changes (i.e., the
sync causes a conflict branch to be created) clean up the blocks as we
did before, by waiting to see the next successul put on the merged
branch.

Issue: KBFS-1261