[KBFS-1190] Handle block get errors gracefully in file/dir removal (#113)
When removing a file, we need to retrieve its top block
in order to account for any other indirect blocks. Also,
when removing a dir, we need to retrieve it in order to
make sure it is empty. In either case, if the block doesn't
exist, go ahead with the removal.

We should still figure out what's causing these bad
pointers, but we shouldn't prevent people from removing
those bad files/dirs.

Also clean up file/dir removal tests a bit.