Bump up connection attempts for Helper tool (macOS)
Patrick experienced a invalid connection after a helper tool upgrade
where the install retried. This is cool because it meant the retry
worked and overcame the flakiness of helper tool upgrades. Though it
seems under this scenario (which is rare) we need to bump up the
retries on the status request after upgrade.