Installer task timeout and stopOnError (macOS) (#3653)
In case of keybase binary being stuck which has been reported a few
times, instead of hanging forever, the task will timeout.

Also instead of installing all the components and then collecting errors,
we'll stop on any error and show a dialog immediately. This ensures a
timely reporting of failure.

We also do a SIGABRT on timed out tasks to get golang stack traces to
help us debug the hang.

I banged on this a whole bunch with binaries that hang.