handle cache misses when loading links from storage
Previously we assumed that the cache would never return nil. This made
some sense, because we first try to load the merkle tip, and we always
cache the tip and links together. However, we've seen this cache miss
happen in the wild (https://github.com/keybase/client/issues/3632), and
it could certainly happen if the client happened to crash at a very
specific time while writing. It makes sense to handle it.