folder_block_ops: fix missing block when switching to indirect
If a series of writes exactly fills up a one-block file, and then a
new write comes in, we transition to an file with a top-level indirect
block, that has two direct file blocks under it -- the original direct
file block, plus a new empty one that receives the write.  However, we
were failing to mark this new block as dirty, because in this case
nCopied == 0 on the first write attempt (because the original block
was full).

In this case, we should dirty that first block, since it needs to be
re-uploaded under a new BlockPointer.

Noticed this while running our benchmark tests.

Issue: KBFS-1364