Fix 'Someone else...' dropdown on mobile
The dropdown calls `this.setState({value})` and then `this._selected()`,
which looks at this.state.value to see what was chosen.  But `setState()`
is asynchronous -- it registers a pending state transition, it's not
atomic.  To wait for setState to complete before inspecting `this.state`,
we can pass a callback as the second argument to `setState()`.