tlf_journal: parallelize flushing put/addRef block ops
This allows the TLF journal to push all block puts in parallel, and
then all addRefs in parallel.  (AddRefs might refer to puts in the
same journal, so we can't parallelize the two types together easily.)
Other operations such as archiveRefs and removeRefs are still done in
serial for now.

Issue: KBFS-1275