[KBFS-1439] Sync blockJournal's in-memory map properly (#295)
Add a tag field to blockRefMap and populate it with
the ordinal. Use that to efficiently remove map
entries on a flush.

Move BlockServerDisk-specific logic out of blockJournal,
and remove more BlockServer-isms in it.

Clean up block_journal_test.go.

Move shutdown logic out of blockJournal and into
tlfJournal and BlockServerDisk.