[KBFS-1535] Keep journal paths to 116 chars or less (#360)
Truncate block and MD paths. Also, instead of
having a device folder and a TLF subfolder, have
just one level of subfolders with a combination
of the truncated device ID and TLF ID.

Store a file in the root of each TLF journal with
the full info, and verify against this info.

Also remove splayed dirs when their contained
files are removed.