Change how we do service driven events.
The old way is actually not supported and a bug that will be fixed in
the next version of redux-saga.

This approach has a couple benefits. It let's us handle things in our
order without failing if there's a slight mismatch. Say we expect A then
B. but we actually get B then A (about the same time).

This can also give us warnings if there is a message the service sent
that we are not explicitly handling.

We can define how many times an incoming call can be made, and error if
that assumption isn't true.