folder_branch_ops: rekeys and CR require an empty journal
Otherwise it indicates there was an error while flushing.

Note that the rekey journal test is removed because with this change
it's impossible to bypass the journal while it's paused to attempt the
rekey.  To make this test work, we'd need to add a lot of machinery to
be able to keep the journal unpaused, but stall the md put of the
journal, which doesn't go through MDOps but directly through MDServer.
I don't think it's worth the trouble to keep the test around.

Issue: KBFS-1589