make ChatThreadConsistencyError an interface
Despite being totally aware that I was risking the tricky nil interface
gotcha by returning a concrete error type, I *still* screwed it up and
broke tests. What I missed was that not only can you coerce to an
interface when you return -- you can also coerce to an interface when
you *assign* if the variable you're assigning to already has an
interface type. And that's exactly what happened with `err` :p

My new strategy is to always define new error types as interfaces, and
to keep the concrete implementation of them totally hidden from callers.
That should make it harder to hit this gotcha.