define a streamable signed-and-encrypted format for chat attachments
The implementation is an encoder and a decoder. Both objects accept any
number of byte slices of any length, and immediately return output bytes
(or errors) to the caller. There are wrapper methods for encoding and
decoding an entire attachment in one shot.

Researching this turned up a bunch of issues related to naive
combinations of encryption and signing. I've tried to address them. See
the (torture-length) design notes comment for the gory details.

This code should have complete line coverage, except for panics that
should never happen.