fix the broken handle_distro_upgrade cronjob for Ubuntu 16.10
Users who have upgraded in-place to Ubuntu 16.10 before we ship this fix
will be stranded, with their /etc/apt/sources.list.d/keybase.list repos
disabled by the upgrade. The cronjob that's supposed to re-enable them
has broken on 16.10, because a config value that was assumed to have a
trailing slash stopped having one.