folder_branch_ops: update latest merged revision on CR
Because CR goes directly to the server when journaling is on (until
KBFS-1637 is in), we can safely update the latest merged revision.  If
we don't, and the TLF isn't accessed through its handle before a
journal flush ends up triggering CR (e.g. because of QR), we could end
up in a situation where head is set, but the latest merged rev is not
(which eventually leads to a panic in the fast-forward code).

Issue: KBFS-1663