actually parse Facebook desktop markup, and unescape '--' properly
Previously we were doing a plain text search for proof text in the
desktop page. That was a little sketchy structurally, but it also ran
into a correctness issue: comments can't contain a "--" character, and
if proofs contained that pair (as it does in a random 1% of our base64
medium sig IDs), we weren't matching it properly.

The simplest possible thing would be to just replace "--" with "-\-\"
before we match. That would *probably* work. However there's a lot of
basic HTML escaping (like &amp;) that's allowed in this context, so
comparing escaped strings char-by-char is an escaping-flexibility
landmine just waiting to explode and break us again.

Instead of looking for escaped text in the page, we should be
*unescaping* text and comparing it to our known string. To do that right
we need to properly structurally parse the page instead of blindly
string searching it. That requires a second parse in the middle, which
is a little complicated, but it works, and it feels much safer in the
long term. (Way less likely to get confused by other text that could
sneak into the page.)