set --auto-forked in run_keybase
When we were auto-restarting the service on Linux after updates, we were
getting errors like this:

    ERROR dial unix /run/user/1000/keybase/keybased.sock: connect: no such file or directory

The problem was that the new client was not detecting that the service
needed to be auto-forked. We need to set the --auto-forked flag when
spawning it in order for the client to know it needs to restart it.

The name of the flag is slightly confusing, since it's more about what
to do in the future than what was done in the past. The long term
solution is probably to dispense with auto-forking entirely and use the
watchdog from Windows, or systemd.