Add basic jest config
We need to provide our own custom babel transformer configuration
because of the layout of our shared directory. Babel looks to parent
directories for .babelrc, and tries to resolve presets/plugins relative
to the .babelrc. We absolutely resolve the babel module locations and
pass this custom config to the babel transformer.