keybase_daemon_rpc: keep the conn alive with CurrentSession calls
OnConnect is called lazily when a reconnection is triggered by an
outgoing RPC, and we need to send HelloIAm calls as soon as we can
after the connection is re-established.  So instead of waiting for
KBFS to make a call, have one going proactively in the background at
all times.

Issue: KBFS-1671