Handle decrypt/verify of pgp messages from non-keybase
This was completely messed up before.  But now, pgp
decrypt, verify on messages encrypted, signed by keys
that keybase doesn't know about are handled correctly.

Also, fixed bug in scankeys where it stored only one
key owner.  The scankeys object can be used for multiple
keys so the owner has to be stored with the key id.
This was causing identifies to happen on the wrong user
in some cases.