file_data: read bytes over multiple levels of indirection
This commit introduces a new set of functions that can read data from
a file over multiple levels of indirection, and fetches the necessary
blocks in parallel whenever possible.  It can copy data into a
pre-made buffer (read()), or can construct a buffer of the correct
size (getBytes()).

If there are holes in the file, it fills the corresponding parts of
the buffer with zeroes.

Use this to reembed block change info, in addition to regular file
reads.

Issue: KBFS-35

fixup