start the gregor connection properly when the network is down
The previous approach would short circuit if the network was down. That
meant that the service would never retry connecting to gregor if the
network came back up. This approach starts the reconnect loop as long as
we've attempted to check the current session with the server. (Though
not if the user is actually logged out.)