file_data: support writes and truncates for >1 indirection levels
Note that this doesn't yet support syncing those multiple levels of
indirection (or CR), so it is turned off for everything but the new
fileData tests right now.

We view the data file as a tree of indirect blocks that ultimately
point to a set of direct leaf blocks at the bottom level.  Whenever
the existing leaf blocks can't accomodate a new write, we append a
leaf block.  If the indirect parent can't accomodate a new block (as
decided by the BlockSplitter), we add a new indirect block at that
level, etc, all the way up to the root.  If the root can't handle a
new block, we add a new level of indirection.

If we are writing into a "hole" in the file that was created by a
previous `truncateExtend` call, the new right block will then have to be
shifted into place by moving around indirect pointers in the interior
of the tree.

When shrinking a file via `truncateShrink`, we truncate indirect
pointer slices and unreference blocks whenever those slices no longer
have any pointers. Note that for now we do not remove levels of
indirection, so the left-most edge of the tree never shrinks.

Issue: KBFS-35