create TimeoutConn/TimeoutDialer and use them in libkb
These wrappers let us specify timeouts that apply for each read and
write, rather than the high-level HTTP timeouts that apply to an entire
request. Previously you could still time out if bytes were trickling in,
but this wrapper should only timeout if there are no bytes over the wire
in the entire timeout interval.

This also refactors libkb.NewClient to keep most of the DefaultTransport
defaults when possible. Previously we were discarding all the defaults
if we did any transport configuration (like Tor mode), which was
unsetting a bunch of sensible limits on connection reuse.

cleanup

even more

working?