folder_branch_ops: don't bother re-registering when MD is too new
We'll never be able to read the MD a folder if we can't understand it.

Might as well just stick to the current version, and if there are
writes, go onto an unmerged branch.  Set a "permanent error" in the
TLF status so we can tell that this happened.

Issue: KBFS-1789