[KBFS-1641] Fix KeyManager test flakes (#619)
The background update task always fetches the head
whenever it receives an update, which triggers a rekey
that races with the rekey-with-prompt timer. Fix this
by, instead of waiting for the first rekey and checking
whether the prompt flag is set, waiting for a rekey
with the prompt flag set. At worst, this'll cause the
test to time out.

Set RekeyWithPromptWaitTime to the individual
test timeout for tests.

Clean up various logs.

Fix bug with runTestsWithMetadataVers.

Also run all MDv3 tests in parallel.