cr: fix handling of previous unmerged resolutionOps
When doing conflict resolution on an unmerged branch that already
contains resolutionOps (e.g., because of local squashes), CR was
incorrectly unreferencing live blocks for files that were
written/synced within a local squash branch.  This is because
`alreadyUpdated` is true, since earlier in `syncBlocks`, these
most-recent file blocks are explicitly added into `updates`. To fix,
we make sure we don't reference anything that is the most recent block
for a chain in the unmerged branch.

A second problem is that `alreadyUpdated` was using update.Unref, but
really what we care about is when the original block for the chain has
been updated (since update chains are collapsed back to the origin
earlier in the function).

Issue: KBFS-1838