folder_block_manager: never give up trying to check revs to delete
It's possible that a user can be throttled so much that they can't
check with the MD server whether a given revision has been written to
the server or not within a reasonable amount of time.  Before, we were
just going ahead and assuming the revision never made it to the
server, and deleting its blocks.  However, if the revision really did
make it to the server, we could be deleting live blocks, and hosing
the TLF!

The reason we gave up after 20 seconds before was because we were
using `getSingleMD`, which returns an error if no revision is found.
However, in this case, knowing that the revision is not found is
important information, and allows us to proceed with deleting the
blocks.

Instead, switch to using getMDRange, and if there's an error, continue
checking the server after a never-ending exponential backoff.  If
there's no error, and no matching MD, we can proceed with deletion
right away.

Issue: KBFS-1730