folder_block_manager: avoid fetch and hash when MD has no data
If an MD has been submitted for deletion that has no serialized
private metadata, we can be sure it was never prepared to go to the
server, and hence we don't need to check with the server before we
clean up its blocks.

This also avoids printing an error when `MakeMdID()` can't find the
private MD.

Noticed while inspecting the logs for other issues.