Fix race in TestCancel in semaphore_test.go (#672)
The spawned goroutine could run before the release
in the main goroutine happens, leading to a race on
the count field of acquireCall. Fix this by doing the
release first before spawning the goroutine.

This only applies for TestCancel, since all other tests
have a successful acquire, which will then have a
count of 0.