folder_branch_ops: set latest merged head when journal conflicts
If we learn about something newer than the base of our journal, we
rightly ignore the updates until CR can resolve the issue.  However,
if we don't set our notion of the latest merged revision, we'll keep
re-registering for the same revision over and over, leading to
ugly throttle messages.

Issue: KBFS-1888