Always stick to the bottom on componentDidUpdate
Sometimes a row will render, while the data stays the same causing us to
no longer be at the bottom. We then don't jump to the bottom because our
condition wasn't met. This makes it so we stick to the bottom if we
started at the bottom on every componentDidUpdate.