folder_branch_ops: on unlink, don't user same path for each ptr
Remove operations of multi-block files can have many unref pointers,
with no indication of which one is the "top" block for the file.
Previously, this code looped through each of them and called
`NodeCache.Unlink` with the same `path` instance each time, just
changing the tail pointer for each call.  However, once one of those
succeeds, the other iterations of the loop will be modifying the same
path that's already been cached in the NodeCache.  This means that if
the user has an open file handle to the file, the next time they read
from it, they may try to read from the wrong block.

Instead, we should break the loop as soon as the node cache takes a
reference to the path, so we don't modify it again.  (We could also
make a new childPath for each iteration of the loop, but that seems
expensive for no good reason.)

Issue: KBFS-1200