folder_branch_ops: always clear the CR branch on a branch change
This fixes a race where, if FBO learns about the new branch when doing
a put (rather than through a onTLFBranchChange callback from the
journal), it wouldn't reset the input state in CR.  This causes CR to
ignore the any inputs with a revision number less than the last one it
processed, and it will cease to perform squashing until the revision
number gets back to where it was before.  But, if there are no new
writes, this will never happen, and journal flushing will remain
paused.

Instead, just clear that input state on every branch change.

Issue: KBFS-1969