tlf_journal: limit squash size by estimate of the unsquashed bytes
If each revision has lots of big blocks, we want to limit the size of
a squash by bytes as well, not just the number of revisions.

This commit does that by tracking an estimate of the number of bytes
that will be in the squash (though it doesn't account for blocks that
will be deleted and not flushed as part of the squash).  If there is
at least one revision with more than a threshold of bytes associated
with it (currently set to 25MB, or about 2 parallel rounds of flushing
max-sized blocks), it will convert it to a squash branch.

If there's only one revision with that many bytes, it can directly be
turned into a squash branch, without the need for CR at all.

Issue: KBFS-1927