block_journal: allow CR to mark squash rev markers as `Ignored`
There was a problem where if a journal containing multiple squashes
encountered a conflict on the server, it would run CR over the entire
journal but not mark the squash revision markers as `Ignored` because
they were already marked as `Unignorable` (since they were from
squashes).  However, a real conflict should be able to override this
Unignorability.

Instead, this gets rid of the concept of `Unignorable`, and instead
marks local squashes as such directly.  This is now safe because we
walk the block journal backwards and quit early while ignoring
revisions, and we used to walk it forwards.

Note that I left the `Unignorable` field itself in case existing
journals have entries with it set, but we should remove it soon.

Issue: KBFS-1992