Replace routeTree persistState parameter with persistChildren route tag
Whether to persist children when navigating away from a route seems to
map consistently to the identity of the route, rather than the context
of the action fire. Moving this to a route tag makes the behavior
consistent with all forms of navigation (e.g. global back button
handler) and simplifies code.