kbfs_ops: eliminate Sync() in favor of SyncAll()
Syncing a single file doesn't make sense in a world where the
directory operation creating the file hasn't been synced yet.  The
simplest solution seems to be to not allow syncing a single file, and
requiring that all operations are synced together no matter what.

So this PR gets rid of Sync(), and makes all of its callers use
SyncAll() instead.

As part of this I got rid of a bunch of mock-based KBFSOps tests that
would have been a pain in the butt to adapt to SyncAll; the stuff they
covered is covered by other, higher-level tests now (those tests
predate the ability to end-to-end FS operations).

Issue: KBFS-2073