Don't consider edits when choosing a snippet
I didn't realize that the service will apply edits to the text messages
in maxMessages itself, so when we're choosing a snippet we can totally
ignore 'edit' messages.

This has the advantage that if someone edits a message that isn't the
latest one, we'll no longer make that old-but-edited message the snippet.