Flatten extra level of login/signup routes
When navigating up (due to back button or iOS left-to-right swipe), we'd
end up on the parent (e.g. loginTab/login/login -> loginTab/login),
which had a defaultSelected value that put us right back on the same
tab. This was confusing the navigation card stack, because it expected
the route to change on back swipe.