dirty_bcache: don't let waitBuf drop below 0
There's a file-write scenario where it's difficult for
`folderBlockOps` to provide correct block accounting to the dirty
bcache:

* A block is dirtied.
* A sync starts.
* The same block is dirtied again by a new write, and that write is deferred.
* The sync fails.
* The block, with the new write, it synced again.

In this case, when the sync fails, it's very hard for `folderBlockOps`
to remember whether the dirty bytes that have been deferred overlap
with dirty bytes that were being synced.  If they do, then we don't
need to subtract them from the dirty bcache waitBuf.  If they don't
overlap, then we do need to subtract them.  But tracking that is hard.

So instead, opt for always subtracting them (as they current code
does), and just don't let that number go negative.

Issue: KBFS-2157