protocol: add UserOrTeamID type and IdentifyLite
KBFS wants to treat users and teams the same in most cases.  So this
adds a new type that can be either a UID or a TeamID, along with
helper methods for it.  It also adds a new `IdentifyLite` RPC that can
identify either a user or a team from an assertion, without needing to
return any keys. (For now, it only supports users.)

Issue: KBFS-2185
Issue: CORE-5273