Remove native dumb sheet flex style
It appears that if the containing Box gets flex: 1 and then the flex
style is removed, the Box will not display. When we specify a style via
parentProps that omits the flex property, it will override the specified
style here and cause this strange flex behavior. Removing the flex rule
here seems like the safe bet; I'm not 100% sure what's going on here,
but we should not expect parentProps.style to specify flex in all cases.