service: infer TLF public from full TLF path
With the addition of team TLFs, using a boolean type for TLFs is
deprecated.  Rekey notifications already provide the full TLF paths
like "/keybase/public/alice", so we can use that to figure out whether
a TLF is public or not.  This is the only user of the
`PublicTopLevelFolder` argument, and I'd like to nuke that argument
entirely in a future PR.

Issue: KBFS-2216