Remove extra focus click component
This was used to make the area after the add button clickable instead of
using onClick on the parent element. This caused a visual glitch when
the input took the full width of the parent, because it would wrap to an
extra blank line. Using onClick turns out to be better anyway, since it
covers the other whitespace and re-focuses the input after 'x' buttons
are clicked.