md_ops: when a TLF is finalized, always allow unverified keys
When trying to verify the MD update that the final update was copied
from, we need to instruct processMetadata to allow unverified keys,
since the last writer user may have been reset.  Do this by altering
the context, because plumbing a boolean through would be too messy.

The included test fails without the rest of the fixes in this PR.

Issue: KBFS-2226