Add tracking to `pgp encrypt`
PGPEncrypt engine now uses NewResolveThenIdentify2WithTrack to verify
recipients, which uses IdentityUI to confirm with local user if the
fetched recipient matches who they want. When user declines, verifyUsers
will fail with "Follow not confirmed by the user." error. When remote
tracking is confirmed, TrackToken engine will be ran with TrackToken
created by ResolveThenIdentify2 engine.

This commit also removes the NewIdentifyMaybeTrackUIProtocol machinery
to decide which IdentifyUI protocol to use basing whether user is logged
in or not. Instead, if PGPEncrypt sees that user is not logged in, it
will pass LocalOnly parameter to ResolveThenIdentify2 engine. This will
give user a chance to review identities.