mdserver_memory: fix locking strategy
There were some functions in this struct, especially `Put()`, that
took the lock multiple times and assumed nothing would change between
the locks.  That led to a panic when md updates were put into the
blocklist out of order.

This refactors things a little to hold the lock over a full function,
and renames functions to make it clear when locks would be needed.

Issue: KBFS-2302