keybase_service_base: call `serviceLoggedIn` on new session
In at least two user reports, the client has gotten into a state where
journaling is not enabled, because the journal thinks it doesn't have
a current UID even though the user is logged in.  This happens when
the KBFS init code can't connect to the service to check the current
session (e.g., because the service is still restarting or down/slow
for some reason).  Later when KBFS does connect to the service, there
is no "logged in" notification (because the user was already logged
in), and so journal-enabling is never reenabled.

This PR executes the logged-in code whenever a new session is noticed,
not just when a new "logged in" notification comes in.

Issue: KBFS-2315