Make PGP commands run ID engine with LocalOnly
We do not post tracking tokens anyway, so make the messaging less-wrong
by setting LocalOnly to true. This will make Identify2 engine not ask
"publicly follow?", only "Is this zzz you wanted?" or "Some proofs
failed; still follow?". The "still follow" here might be misleading
because it's about local tracking. There is still an issue where we
don't check for confirmation result here (e.g. pressing "NO" will still
encrypt and return armored message), but it is a cache related defect in
Identify2 engine which we decided not to fix right now.