Attempt to fix "message stuck in pending state" bugs
In the olden days, pending messages worked like this:

* we send a message

* when it succeeds or fails, we get an incoming new chat activity
  message that tells us what happens

* we update the pending message in our store to match the new state

Now, though, something else can happen:

* we send a message

* we're offline for a bit

* we come back online

* gregor tells us that our understanding of the thread is stale and
  we should get newer messages

* we do so

* the new thread includes the sent version of our previously pending message

In this version, we still have our pending message in our store, because
nothing updated it when the new message came in.

This PR does that by deleting a pending message if we receive a message
thread containing a message that matches its outboxID, and then pushing
the new sent message in its place.