add IsDirectlyRevoked to the TypedChainLink interface
TrackChainLinks consider themselves "IsRevoked=true" when an
UntrackChainLink exists for the same user. We don't really want to test
for that behavior in the sigchain test vectors, though, since it has
more to do with the application level than with the sigchain itself. So
far we haven't had a test case with an untrack link, so this discrepancy
never came up, but we're about to add Max's sigchain.