kbfsgit: support simple fetches
This may not cover all corner cases, but it works for simple cases.
It sets the local repo as a "remote" in the bare repo, and pushes from
the bare repo into this "remote".  go-git requires pushing to a ref
name, but `git` gets upset if you push into the refname it's
expecting, so we have to use a temporary name and then delete it
afterwards.

Issue: KBFS-2353