kbfsgit: interrupt on `git` control-C when stdin is closed
The only reliable way to tell if the calling `git` process has been
ctrl-c'd is if `stdin` is killed.  If that happens before the remote
helper gets a blank line command, then the remote helper should exit
immediately.

Accomplish this by processing individual commands in a separate
goroutine (via a channel), and interrupt that goroutine when EOF is
read from stdin.

The added test hangs without the rest of the commit.

Issue: KBFS-2408