Don't blow away architecture specifiers
When using multiarch on Debian, apt will (by default) produce a warning
for each and every repository that does not contain one or more of the
architectures that was added to the configuration with "dpkg
--add-architecture". However, this warning can be silenced by adding an
architecture specifier to an apt line, which (in the case of
keybase) could look like so:

    deb [arch=amd64,i386] http://prerelease.keybase.io/deb stable main

The grep in the current version of the cron job does not match that
(optional) architecture specifier, therefore does not match the line,
and therefore blows it away. This is annoying.

Fix by changing the grep line to allow for it.