kbfsgit: initialize repo with a packed-refs file if many refs
If the local repo has non-branch, non-tag refs, and `git push --all`
is used, then we won't be able to use `pushAll` because not all refs
will be pushed (despite the command line flag saying otherwise).  But
there could be lots of branches and tags, and we would still benefit
from having a packed-refs file.

So if we are initializing an empty repo, and there are more than a
handful of branches/tags to push, ask go-git to make a packed-refs
file for us, instead of individual ref files.

Issue: KBFS-2464