logger: don't build on linux at all
We're going to be using dup3 system call on linux so we create a new
file.

logger: Introduce redirect logger for linux

Because aarch64 doesn't support dup2, we need to use dup3 which is
supported on Linux (but not *nix) so we create a file for linux and only
build it when compiling for linux.

Fix flag to be 0