Delete entry if entry's key is no longer valid.
This can happen if the secure lock screen is disabled (i.e., reconfigured to None, Swipe or other mode which does not authenticate the user) or when the secure lock screen is forcibly reset (e.g., by Device Admin).

We delete it, and let the setupKeyStore create a new key entry.