keybase_daemon_rpc: put short timeouts on service pings
The whole point of the pings is to force quick reconnects of KBFS to
the service, if the service restarts.  However, if a ping itself gets
"stuck" in a connection to a daemon that has restarted, we need to
make sure that pings gets canceled so the next one will be sent.  So
use 1 second timeout for that.

Issue: KBFS-2478