journal_server: error, instead of panic, if no handle given
We require a handle to be passed in when enabling a journal for a team
TLF, since otherwise we'd have to talk to the (untrusted) server to
look up the `chargedTo` ID for the folder.  Previously we panicked
when the caller didn't provide a handle.

However there are legitimate callers, like
`folderBranchOps.getJournalPredecessorRevision()`, which might try to
inadvertently enable a journal without passing in a handle, if they
happen to be called before some other `folderBranchOps` function.
These callers generally expect an error if the journal is not
enablable.  So let's turn the panic into an error instead.

Issue: KBFS-2563