folder_branch_ops: avoid data race in `Lookup()`
If `Lookup` returns early, it could write to the return param `node`.
But the internal goroutine might still be running, which also tried to
write to the `node` param.  This leads to a possible (but really rare)
data race.

Instead, have the goroutine write to a temp variable, which is only
returned when the `Lookup` is successful.

Issue: KBFS-2579