fix a nil context getting passed to OutputSignatureSuccessNonKeybase
The original issue (CORE-6440) was that `keybase pgp verify` wasn't
printing anything for valid messages signed by revoked keys. This turned
out to be caused by a nil context (`ctx.NetContext`) that was getting
passed into an RPC call. That caused the rpc library to short-circuit
immediately with `errors.New("No Context provided for this call")`,
however we were dropping the error. As part of fixing the nil context,
make sure we propagate errors here.

TODO?: There are many other places in the codebase referencing
`ctx.NetContext` directly, as this code did. Are all of them suspect?