libfs: a quick way to get a handle, without getting the TLF ID
When listing favorites, it would be a bummer for performance if the
ReadDir had to wait for (at least) two RTTs to the servers, for
checking on implicit teams, and fetching the TLF ID.  Luckily, we
don't need to know these until someone actually tries to access data
within the folder.

So add a method for doing faking out these time-consuming calls during
favorites listing, and then updating the handle later when it is
really needed.

Issue: KBFS-2598