prefer `restart` to `start` with systemd
The problem with `start` in `run_keybase` is that it races against
`kill`. There's a window where the kill signal has been sent, but the
service has not yet shut down, and `start` is still a no-op in that
window. I don't know if this would happen with SIGKILL, but it's
certainly an issue with SIGTERM, since we run a handler.