make the Debian install/cronjob insert out PGP key every time
This is the crucial change that will let us update the code signing key
on Debian before it expires in 5 days. We added the updated key body in
a0c9ba3.

I tested this by un-suppressing the output of the apt-key command in the
cronjob and manually running it repeatedly:

1) The key does get updated. (This happens immediately after the package
   update, because the postinstall script runs the cronjob.)
2) After the first time, the /etc/apt/trusted.gpg keyring doesn't get
   modified. So we don't have to worry about e.g. the file growing over
   time.
3) Corrupting the signing key does make apt-key return an error.