MapUIDsReturnMap change, fix implicit admin Active flag
Made it possible for MapUIDsReturnMap to return partial data and error,
similar to how MapUIDsToUsernamePackages works. Adapted teams/list code
to be aware of that.

Fixed invalid flagging of members as not active when a member is an
implicit admin without real membership in a team. Implicit admins are
server trust: ListAll code does not walk through team chain to find such
user's UV in one of ancestor teams, so EldestSeqno stays at default 0.