kbpki_client: handle errors during writer checks for public iteams
The service returns a `libkb.NotFoundError` when trying to load a team
if you're not a member of that team.  We don't want that error to
propagate up to the rest of KBFS, since we might be checking
writership in order to set permissions on a file in a public TLF, for
example.  So ignore those errors in the public iteam case, and avoid
trying to load the team at all if we're not logged in.