Stylistic/comment change in handleSBSSingle (#10097)
This function takes untrustedInviteeFromGregor as parameter, and does
the following:
1) Load team,
2) look for invite with id=untrustedInviteeFromGregor.InviteID,
3) verify that untrustedInviteeFromGregor.UV is the real invitee from
   on-chain invite and that we can bring them in,
4) Add untrustedInviteeFromGregor.UV to the team.

In step 4, using untrustedInviteeFromGregor might seem worrying, as if
we were using untrusted from the server. But we've just verified
everything in 3.

Make the code a bit more clear about what's going on, even if someone
just glances at the code.