libgit: fix autogit updating when source repo is read-only
When go-git calls `OpenFile` on an object file, they expect to get
back a not-exists error if the object doesn't exist.  However, if the
object's parent directory didn't exist, we were trying to create it
automatically (which go-git _also_ expects to happen, *if* the file is
successfully created).  If the source repo is read-only (say, someone
else's public directory), this doesn't work because `OpenFile`
returned a `WriteAccessError` instead of a non-exists error.

This commit fixes that issue by returning a not-exists error in that
case.  It also splits up the autogit update tests to have a separate
read-only test, and makes that read-only test repack the objects and
delete the object directories in order to repro the problem.  The test
fails without this fix.