Factor out subcomponents of info panel (#10219)
This is in preparation for merging all the info panels into one component
that pushes sub-components onto a virtualized list.

Also move the mute row into the Notifications component and make it always
have the same label, per the design. This is also in preparation for making
a checked mute box disable/hide the other notification components.