Transactional AddMember (#10278)
This pull request adds a new file teams/transactions.go that
introduces AddMemberTx struct with methods. It can be used to add one
or more member to a team that may or may not need to post multiple
signatures to achieve the task, e.g.:

1) first cancel existing keybase-type invite, then add member with
ChangeMembership (two sigs),
2) first remove existing member old UV, then add new member UV (one
sig),
3) first remove existing member old UV, then add new keybase-type
invite (two sigs).

AddMemberTx has payloads list of values that are either *SCTeamInvites
or *keybase1.TeamChangeReq. When adding members, functions will either
find existing payload of needed type, or add a new one. When tx.Post
is called, payloads are transformed to SCTeamSections, and then chain
of signatures is created to be posted using sig/multi. No serverside
changes were needed to get this to work.